追加: 番組統計エリアを追加
